LayoutTests/platform/blackberry/fast/dom/serialize-nodes-expected.txt

ALERT: foo:name is one and should be one
ALERT: bar:name is two and should be two
ALERT: name is three and should be three
ALERT: node is child and should be child
ALERT: <foo:root xmlns:foo="urn:foo-ns" foo:type="test"><child foo:name="one" bar:name="two" xmlns:bar="urn:bar-ns" name="three"/></foo:root>
<div id="input">
    <div>
        <foo:node xmlns="http://baz.com" foo:name="foo_name" bar:name="bar_name">
            <node foo:name="foo_name">
                <bar:node xmlns:bar="http://bar2.com">
            </bar:node></node>
        </foo:node>
    </div>
    <bar:node>
        <br>
    </bar:node>
</div>
